Try out setting predict endpoint from websocket request
